Upgrade to spring boot 3.5 and axon 4.9

+++ backend/services/project/application/src/main/resources/db/migration/postgres/V3__QUERY_MODEL_TABLE.sql
@@ -54,7 +54,7 @@ create table tasks
     name        varchar(255) not null,
     project_id  varchar(255) not null,
     start_date  date         not null,
-    status      int4         not null,
+    status      smallint     not null,
     version     int8         not null,
     primary key (identifier)
 );
